Update poster.tex
I have changed a few things throughout. I note that the table 1 is unclear. Is HbA1c used to define prediabetes and diabetes. I presume in one case HbA1c<6.5 and in the other it is just >6.5.
Also I would have done the analysis also for the ones with missing, just to see if there might be aby bias due to missing data.